Rename Vanilla\Community\Events\UserEvent to Vanilla\Dashboard\Events\UserEvent
