+++ tests/Models/UserModelTest.php
@@ -9,7 +9,7 @@ namespace VanillaTests\Models;
 use VanillaTests\SharedBootstrapTestCase;
 use VanillaTests\SiteTestTrait;
 use Garden\EventManager;
-use Vanilla\Community\Events\UserEvent;
+use Vanilla\Dashboard\Events\UserEvent;
 
 /**
  * Test {@link UserModel}.
